fix: move speed metrics to result sets
